+++ compose-material3-pickers/src/main/java/com/rtamayo/compose_material3_pickers/date/models/Month.kt
@@ -1,4 +1,4 @@
-package com.rtamayo.compose_material3_pickers.datepicker.models
+package com.rtamayo.compose_material3_pickers.date.models
 
 import java.time.DayOfWeek
 import java.time.Month
