+++ compose-material3-pickers/src/main/java/com/rtamayo/compose_material3_pickers/date/utils/DateIterator.kt
@@ -1,4 +1,4 @@
-package com.rtamayo.compose_material3_pickers.datepicker.utils
+package com.rtamayo.compose_material3_pickers.date.utils
 
 import java.time.LocalDate
 import java.time.temporal.ChronoUnit
@@ -6,8 +6,8 @@ import java.util.stream.Stream
 import kotlin.streams.toList
 
 class DateIterator(
-    val startDate: LocalDate,
-    val endDate: LocalDate
+    private val startDate: LocalDate,
+    private val endDate: LocalDate
 ): Iterable<LocalDate> {
     override fun iterator(): Iterator<LocalDate> {
         return stream().iterator()
